hand camera support for policy learning, code refactoring

+++ scripts/bash/run_pretrain.sh
@@ -7,7 +7,9 @@ fi
 
 if [ $1 == "test" ]; then
     echo "Running test run -> removing existing experiment artifacts"
-    rm -r /home/fmorro/INEST-MANISKILL/experiments/pretrain/test
+    if [ -d "/home/fmorro/INEST-MANISKILL/experiments/pretrain/test" ]; then
+        rm -r /home/fmorro/INEST-MANISKILL/experiments/pretrain/test
+    fi
 fi
 
 TMPDIR=/home/fmorro/tmp MPLBACKEND=Agg python /home/fmorro/INEST-MANISKILL/scripts/pretrain.py \
